requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
autoflake==0.7
autopep8==1.3.2
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
coverage==4.4.1           # via coveralls, pytest-cov
coveralls==1.1
django-autoslug==1.9.3
django-easy-select2==1.3.4
django-enumfields==0.9.0
django-environ==0.4.3
django-filter==1.0.4
django-parler-rest==1.4.2
django-parler==1.8        # via django-parler-rest
django==1.11.3
djangorestframework-gis==0.11.2
djangorestframework==3.6.3
docopt==0.6.2             # via coveralls
factory-boy==2.8.1        # via pytest-factoryboy
faker==0.7.17             # via factory-boy
first==2.0.1              # via pip-tools
flake8-isort==2.2.2
flake8-polyfill==1.0.1    # via flake8-isort
flake8==3.3.0
httpio==0.1.3
idna==2.5                 # via requests
inflection==0.3.1         # via pytest-factoryboy
isort==4.2.15             # via flake8-isort
jsonschema==2.6.0
lxml==3.8.0
mccabe==0.6.1             # via flake8
pip-tools==1.9.0
psycopg2==2.7.1
py==1.4.34                # via pytest
pycodestyle==2.3.1        # via autopep8, flake8
pyflakes==1.5.0           # via autoflake, flake8
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-factoryboy==1.3.1
pytest==3.1.3             # via pytest-cov, pytest-django, pytest-factoryboy
python-dateutil==2.6.1    # via faker
pytz==2017.2
pyyaml==3.12
requests==2.18.1          # via coveralls, httpio
six==1.10.0               # via django-environ, faker, pip-tools, python-dateutil
testfixtures==5.1.1       # via flake8-isort
urllib3==1.21.1           # via requests
